Rename api to http
